requirements.txt
requests
coverage
lxml==3.4.4
mock
nose
python-dateutil
pytz
future
